refactor to rename pull_request to hosting_service and apply SRP
